// src/app/modules/hcm/shared/models/masters/accounting-item-segment.ts
export class AccountingItemSegment {

    idAccountingItem: number =-1;
    idAccountingItemDetail: number =-1;
    idSegmentType: number =-1;
    identifier: string ="";
    ordinal: number =-1;
    description: string ="";
    active: boolean = false;
    createdByUser: string = "";
    createdByUserId: number = -1;
    updatedByUser: string = "";
    updatedByUserId: number = -1;
}